Fix conversion errors by making conversions explicit.
svn path=/trunk/kdesupport/strigi/; revision=1059956

+++ src/streams/dostime.cpp
@@ -47,12 +47,12 @@ dos2unixtime (unsigned long dostime)
   /* Call localtime to initialize timezone in TIME.  */
   ltime = *localtime (&now);
 
-  ltime.tm_year = (dostime >> 25) + 80;
-  ltime.tm_mon = ((dostime >> 21) & 0x0f) - 1;
-  ltime.tm_mday = (dostime >> 16) & 0x1f;
-  ltime.tm_hour = (dostime >> 11) & 0x0f;
-  ltime.tm_min = (dostime >> 5) & 0x3f;
-  ltime.tm_sec = (dostime & 0x1f) << 1;
+  ltime.tm_year = (int)(dostime >> 25) + 80;
+  ltime.tm_mon = ((int)(dostime >> 21) & 0x0f) - 1;
+  ltime.tm_mday = (int)(dostime >> 16) & 0x1f;
+  ltime.tm_hour = (int)(dostime >> 11) & 0x0f;
+  ltime.tm_min = (int)(dostime >> 5) & 0x3f;
+  ltime.tm_sec = (int)(dostime & 0x1f) << 1;
 
   ltime.tm_wday = -1;
   ltime.tm_yday = -1;
